Fertig

+++ MyApp/Controllers/Users.php
@@ -90,13 +90,17 @@ class Users extends Framework
         // Alle Tracks des Users holen.
         $tracks = $this->Tracks_Model->get_tracks_on_user($user_id);
 
-        // Alle Tracks des Users löschen
-        foreach($tracks as $track) {
-            $this->Tracks_Model->delete_track($user_id, $track["track_id"]);
-        }
+        // Nur wenn der Benutzer eigentümer ist
+        if($user_id == $this->modules->Session->get("user_id")) {
 
-        // User löschen
-        $this->Users_Model->delete_user($user_id);
+            // Alle Tracks des Users löschen
+            foreach ($tracks as $track) {
+                $this->Tracks_Model->delete_track($user_id, $track["track_id"]);
+            }
+
+            // User löschen
+            $this->Users_Model->delete_user($user_id);
+        }
 
     }
 }
